Merged PR 2038: Updated Dockerfile
Updated Dockerfile. Since the toolkit-execute file is edited, the lines following the one that copy toolkit-execute to the docker image will be re-compiled, which leads to the time for builds exceeding 60 minutes. Can the administrators allow the compilation time exceeding 60 minutes? I tried several times but everytime failed because of exceeding 60 minutes.